Refactor some code to make logic simpler + restyle table buttons + rename to ManageColumns
